fix(gui): handle invalid data when copying belongs to cell data
